Fixed datatype mismatch in comparison

+++ update/Default.aspx.cs
@@ -339,7 +339,7 @@ public partial class update_Default : System.Web.UI.Page
 	{
 		if (enforceBusinessRules)
 		{
-			int newHeight = int.Parse(txtAntennaHeight.Text);
+			Double newHeight = Double.Parse(txtAntennaHeight.Text);
 			Double allowedHeight = Double.Parse(repeater.CoordinatedAntennaHeight) + 15.24;
 
 			if (newHeight <= allowedHeight)
